modified some file and can calculate on F_q.

+++ include/util.h
@@ -2,6 +2,9 @@
 
 #include <stdio.h>
 
+#include "gmp.h"
+#include "gmpxx.h"
+
 // GCD
 int gcd(int a,int b);
 // Extended Euclid's Algorithm with GCD.
@@ -10,4 +13,4 @@ void xeuclid(const int a,const int b,int *u,int *v,int *g);
 void xeuclid(const int a,const int b,int *u,int *v);
 // Calculate the inverse in F_m.
 // If we get -1, there doesn't exist inverse in F_m.
-int inv_mod(int a,int m);
+long long inv_mod(long long a,long long m);
